rename app provider to chat provider
